Extract linked-PR integration test to its own module (issue #187)
Move test_parse_issue_search_json_populates_linked_pr_numbers out of
issues.rs into a new tests_linked_prs.rs module. The 60-line fixture
test pushed issues.rs past the 1000-line source-size hard limit; the
sibling-module pattern (matching state_reason.rs / tests_filters.rs)
keeps each file under the policy.
